refactor: split workoutPresetForm into custom hook
