+++ SparkyFitnessFrontend/src/pages/Exercises/WorkoutPresetForm.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useCallback } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import {
   Dialog,
@@ -13,53 +13,15 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
 import { Switch } from '@/components/ui/switch';
-import type {
-  WorkoutPreset,
-  WorkoutPresetExercise,
-  WorkoutPresetSet,
-} from '@/types/workout';
+import type { WorkoutPreset } from '@/types/workout';
 import AddExerciseDialog from './AddExerciseDialog';
-import ExerciseHistoryDisplay from '../../components/ExerciseHistoryDisplay';
-import {
-  Plus,
-  X,
-  Repeat,
-  Timer,
-  GripVertical,
-  Copy,
-  Dumbbell,
-  Hourglass,
-} from 'lucide-react';
+import { Plus } from 'lucide-react';
 import { usePreferences } from '@/contexts/PreferencesContext';
-import { useToast } from '@/hooks/use-toast';
-import { generateClientId } from '@/utils/generateClientId';
-import { debug } from '@/utils/logging';
-import {
-  DndContext,
-  closestCenter,
-  KeyboardSensor,
-  PointerSensor,
-  useSensor,
-  useSensors,
-  type DragEndEvent,
-} from '@dnd-kit/core';
-import {
-  SortableContext,
-  sortableKeyboardCoordinates,
-  arrayMove,
-  useSortable,
-} from '@dnd-kit/sortable';
-import { CSS } from '@dnd-kit/utilities';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-import { Exercise } from '@/types/exercises';
-import { TFunction } from 'i18next';
-import { UnitInput } from '@/components/ui/UnitInput';
+import { DndContext, closestCenter } from '@dnd-kit/core';
+import { SortableContext } from '@dnd-kit/sortable';
+
+import { SortableExerciseItem } from './SortableExerciseItem';
+import { useWorkoutPresetForm } from '@/hooks/Exercises/useWorkoutPresetForm';
 
 interface WorkoutPresetFormProps {
   isOpen: boolean;
@@ -70,322 +32,6 @@ interface WorkoutPresetFormProps {
   initialPreset?: WorkoutPreset | null;
 }
 
-const SortableSetItem = React.memo(
-  ({
-    t,
-    set,
-    exerciseIndex,
-    setIndex,
-    onSetChange,
-    onDuplicateSet,
-    onRemoveSet,
-    weightUnit,
-  }: {
-    t: TFunction;
-    set: WorkoutPresetSet;
-    exerciseIndex: number;
-    setIndex: number;
-    onSetChange: (
-      exerciseIndex: number,
-      setIndex: number,
-      field: keyof WorkoutPresetSet,
-      value: WorkoutPresetSet[keyof WorkoutPresetSet]
-    ) => void;
-    onDuplicateSet: (exerciseIndex: number, setIndex: number) => void;
-    onRemoveSet: (exerciseIndex: number, setIndex: number) => void;
-    weightUnit: string;
-  }) => {
-    const { attributes, listeners, setNodeRef, transform, transition } =
-      useSortable({ id: set.id! });
-
-    const style = {
-      transform: CSS.Transform.toString(transform),
-      transition,
-    };
-
-    return (
-      <div
-        ref={setNodeRef}
-        style={style}
-        className="flex flex-col space-y-2"
-        {...attributes}
-      >
-        <div className="flex items-center space-x-2">
-          <div {...listeners}>
-            <GripVertical className="h-5 w-5 text-muted-foreground cursor-grab" />
-          </div>
-          <div className="grid grid-cols-1 md:grid-cols-8 gap-2 flex-grow items-center">
-            <div className="md:col-span-1">
-              <Label>{t('workoutPresetForm.setLabel', 'Set')}</Label>
-              <p className="font-medium p-2">{set.set_number}</p>
-            </div>
-            <div className="md:col-span-2">
-              <Label>{t('workoutPresetForm.typeLabel', 'Type')}</Label>
-              <Select
-                value={set.set_type || undefined}
-                onValueChange={(value) =>
-                  onSetChange(exerciseIndex, setIndex, 'set_type', value)
-                }
-              >
-                <SelectTrigger>
-                  <SelectValue
-                    placeholder={t(
-                      'workoutPresetForm.setTypePlaceholder',
-                      'Set Type'
-                    )}
-                  />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="Working Set">
-                    {t('workoutPresetForm.workingSet', 'Working Set')}
-                  </SelectItem>
-                  <SelectItem value="Warm-up">
-                    {t('workoutPresetForm.warmUp', 'Warm-up')}
-                  </SelectItem>
-                  <SelectItem value="Drop Set">
-                    {t('workoutPresetForm.dropSet', 'Drop Set')}
-                  </SelectItem>
-                  <SelectItem value="Failure">
-                    {t('workoutPresetForm.failure', 'Failure')}
-                  </SelectItem>
-                  <SelectItem value="AMRAP">
-                    {t('workoutPresetForm.amrap', 'AMRAP')}
-                  </SelectItem>
-                  <SelectItem value="Back-off">
-                    {t('workoutPresetForm.backOff', 'Back-off')}
-                  </SelectItem>
-                  <SelectItem value="Rest-Pause">
-                    {t('workoutPresetForm.restPause', 'Rest-Pause')}
-                  </SelectItem>
-                  <SelectItem value="Cluster">
-                    {t('workoutPresetForm.cluster', 'Cluster')}
-                  </SelectItem>
-                  <SelectItem value="Technique">
-                    {t('workoutPresetForm.technique', 'Technique')}
-                  </SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="md:col-span-1">
-              <Label
-                htmlFor={`reps-${exerciseIndex}-${set.id}`}
-                className="flex items-center"
-              >
-                <Repeat className="h-4 w-4 mr-1" style={{ color: '#3b82f6' }} />{' '}
-                {t('workoutPresetForm.repsLabel', 'Reps')}
-              </Label>
-              <Input
-                id={`reps-${exerciseIndex}-${set.id}`}
-                type="number"
-                value={set.reps ?? ''}
-                onChange={(e) =>
-                  onSetChange(
-                    exerciseIndex,
-                    setIndex,
-                    'reps',
-                    Number(e.target.value)
-                  )
-                }
-              />
-            </div>
-            <div className="md:col-span-1">
-              <Label
-                htmlFor={`weight-${exerciseIndex}-${set.id}`}
-                className="flex items-center"
-              >
-                <Dumbbell
-                  className="h-4 w-4 mr-1"
-                  style={{ color: '#ef4444' }}
-                />{' '}
-                {t('workoutPresetForm.weightLabel', 'Weight')} ({weightUnit})
-              </Label>
-              <UnitInput
-                id={`weight-${exerciseIndex}-${set.id}`}
-                type="weight"
-                unit={weightUnit}
-                value={set.weight ?? 0}
-                onChange={(metricValue) =>
-                  onSetChange(exerciseIndex, setIndex, 'weight', metricValue)
-                }
-              />
-            </div>
-            <div className="md:col-span-1">
-              <Label
-                htmlFor={`duration-${exerciseIndex}-${set.id}`}
-                className="flex items-center"
-              >
-                <Hourglass
-                  className="h-4 w-4 mr-1"
-                  style={{ color: '#f97316' }}
-                />{' '}
-                {t('workoutPresetForm.durationLabel', 'Duration (min)')}
-              </Label>{' '}
-              <Input
-                id={`duration-${exerciseIndex}-${set.id}`}
-                type="number"
-                value={set.duration ?? ''}
-                onChange={(e) =>
-                  onSetChange(
-                    exerciseIndex,
-                    setIndex,
-                    'duration',
-                    Number(e.target.value)
-                  )
-                }
-              />
-            </div>
-            <div className="md:col-span-1">
-              <Label
-                htmlFor={`rest-${exerciseIndex}-${set.id}`}
-                className="flex items-center"
-              >
-                <Timer className="h-4 w-4 mr-1" style={{ color: '#8b5cf6' }} />{' '}
-                {t('workoutPresetForm.restLabel', 'Rest (s)')}
-              </Label>
-              <Input
-                id={`rest-${exerciseIndex}-${set.id}`}
-                type="number"
-                value={set.rest_time ?? ''}
-                onChange={(e) =>
-                  onSetChange(
-                    exerciseIndex,
-                    setIndex,
-                    'rest_time',
-                    Number(e.target.value)
-                  )
-                }
-              />
-            </div>
-            <div className="flex items-center space-x-1">
-              <Button
-                variant="ghost"
-                size="icon"
-                onClick={() => onDuplicateSet(exerciseIndex, setIndex)}
-              >
-                <Copy className="h-4 w-4" />
-              </Button>
-              <Button
-                variant="ghost"
-                size="icon"
-                onClick={() => onRemoveSet(exerciseIndex, setIndex)}
-              >
-                <X className="h-4 w-4" />
-              </Button>
-            </div>
-          </div>
-        </div>
-        <div className="pl-8">
-          <Label htmlFor={`notes-${exerciseIndex}-${set.id}`}>
-            {t('workoutPresetForm.notesLabel', 'Notes')}
-          </Label>
-          <Textarea
-            id={`notes-${exerciseIndex}-${set.id}`}
-            value={set.notes ?? ''}
-            onChange={(e) =>
-              onSetChange(exerciseIndex, setIndex, 'notes', e.target.value)
-            }
-          />
-        </div>
-      </div>
-    );
-  }
-);
-
-const SortableExerciseItem = React.memo(
-  ({
-    ex,
-    exerciseIndex,
-    handleRemoveExercise,
-    handleSetChange,
-    handleDuplicateSet,
-    handleRemoveSet,
-    handleAddSet,
-    weightUnit,
-    t,
-  }: {
-    ex: WorkoutPresetExercise;
-    exerciseIndex: number;
-    handleRemoveExercise: (index: number) => void;
-    handleSetChange: (
-      exerciseIndex: number,
-      setIndex: number,
-      field: keyof WorkoutPresetSet,
-      value: WorkoutPresetSet[keyof WorkoutPresetSet]
-    ) => void;
-    handleDuplicateSet: (exerciseIndex: number, setIndex: number) => void;
-    handleRemoveSet: (exerciseIndex: number, setIndex: number) => void;
-    handleAddSet: (exerciseIndex: number) => void;
-    weightUnit: string;
-    t: TFunction;
-  }) => {
-    const {
-      attributes,
-      listeners,
-      setNodeRef,
-      transform,
-      transition,
-      isDragging,
-    } = useSortable({ id: ex.id! });
-
-    const style = {
-      transform: CSS.Transform.toString(transform),
-      transition,
-      opacity: isDragging ? 0.5 : 1,
-    };
-
-    return (
-      <div
-        ref={setNodeRef}
-        style={style}
-        className="border p-4 rounded-md space-y-4 bg-card"
-        {...attributes}
-      >
-        <div className="flex justify-between items-center">
-          <div className="flex items-center space-x-2">
-            <div {...listeners}>
-              <GripVertical className="h-5 w-5 text-muted-foreground cursor-grab" />
-            </div>
-            <h4 className="font-semibold">{ex.exercise_name}</h4>
-          </div>
-          <Button
-            variant="ghost"
-            size="icon"
-            onClick={() => handleRemoveExercise(exerciseIndex)}
-          >
-            <X className="h-4 w-4" />
-          </Button>
-        </div>
-        <SortableContext items={ex.sets.map((s) => s.id!)}>
-          <div className="space-y-2">
-            {ex.sets.map((set, setIndex) => (
-              <SortableSetItem
-                key={set.id}
-                t={t}
-                set={set}
-                exerciseIndex={exerciseIndex}
-                setIndex={setIndex}
-                onSetChange={handleSetChange}
-                onDuplicateSet={handleDuplicateSet}
-                onRemoveSet={handleRemoveSet}
-                weightUnit={weightUnit}
-              />
-            ))}
-          </div>
-        </SortableContext>
-        <Button
-          type="button"
-          variant="outline"
-          onClick={() => handleAddSet(exerciseIndex)}
-        >
-          <Plus className="h-4 w-4 mr-2" />{' '}
-          {t('workoutPresetForm.addSetButton', 'Add Set')}
-        </Button>
-        <ExerciseHistoryDisplay exerciseId={ex.exercise_id} />
-      </div>
-    );
-  }
-);
-
 const WorkoutPresetForm: React.FC<WorkoutPresetFormProps> = ({
   isOpen,
   onClose,
@@ -393,264 +39,34 @@ const WorkoutPresetForm: React.FC<WorkoutPresetFormProps> = ({
   initialPreset,
 }) => {
   const { t } = useTranslation();
-  const { loggingLevel, weightUnit } = usePreferences();
-  const { toast } = useToast();
-  const [name, setName] = useState(initialPreset?.name || '');
-  const [description, setDescription] = useState(
-    initialPreset?.description || ''
-  );
-  const [isPublic, setIsPublic] = useState(initialPreset?.is_public ?? false);
-  const [exercises, setExercises] = useState<WorkoutPresetExercise[]>(() => {
-    return (
-      initialPreset?.exercises.map((ex) => ({
-        ...ex,
-        id: ex.id ? String(ex.id) : generateClientId(),
-        sets: ex.sets.map((set) => ({
-          ...set,
-          id: set.id ? String(set.id) : generateClientId(),
-          weight: Number(set.weight) || 0, // Keep metric (kg)
-        })),
-      })) || []
-    );
-  });
-  const [isAddExerciseDialogOpen, setIsAddExerciseDialogOpen] = useState(false);
+  const { weightUnit } = usePreferences();
 
-  const handleAddExercise = (exercise: Exercise | undefined) => {
-    if (exercise) {
-      const newExercise: WorkoutPresetExercise = {
-        id: generateClientId(), // Stable ID for DND
-        exercise_id: exercise.id,
-        exercise_name: exercise.name,
-        image_url:
-          exercise.images && exercise.images.length > 0
-            ? exercise.images[0]
-            : '',
-        exercise: exercise,
-        sets: [
-          {
-            id: generateClientId(),
-            set_number: 1,
-            set_type: 'Working Set',
-            reps: 10,
-            weight: 0,
-          },
-        ],
-      };
-      setExercises((prev) => [...prev, newExercise]);
-    }
-    setIsAddExerciseDialogOpen(false);
-  };
-
-  const handleRemoveExercise = (index: number) => {
-    setExercises((prev) => prev.filter((_, i) => i !== index));
-  };
-
-  const handleSetChange = useCallback(
-    (
-      exerciseIndex: number,
-      setIndex: number,
-      field: keyof WorkoutPresetSet,
-      value: WorkoutPresetSet[keyof WorkoutPresetSet]
-    ) => {
-      setExercises((prev) =>
-        prev.map((exercise, eIndex) => {
-          if (eIndex !== exerciseIndex) {
-            return exercise;
-          }
-          return {
-            ...exercise,
-            sets: exercise.sets.map((set, sIndex) => {
-              if (sIndex !== setIndex) {
-                return set;
-              }
-              return { ...set, [field]: value };
-            }),
-          };
-        })
-      );
-    },
-    []
-  );
-
-  const handleAddSet = useCallback((exerciseIndex: number) => {
-    setExercises((prev) =>
-      prev.map((exercise, eIndex) => {
-        if (eIndex !== exerciseIndex) {
-          return exercise;
-        }
-        const lastSet = exercise.sets[exercise.sets.length - 1];
-        if (!lastSet) {
-          return exercise;
-        }
-        const newSet: WorkoutPresetSet = {
-          ...lastSet,
-          id: generateClientId(),
-          set_number: exercise.sets.length + 1,
-        };
-        return {
-          ...exercise,
-          sets: [...exercise.sets, newSet],
-        };
-      })
-    );
-  }, []);
-
-  const handleDuplicateSet = useCallback(
-    (exerciseIndex: number, setIndex: number) => {
-      setExercises((prev) =>
-        prev.map((exercise, eIndex) => {
-          if (eIndex !== exerciseIndex) {
-            return exercise;
-          }
-          const sets = exercise.sets;
-          const setToDuplicate = sets[setIndex];
-          if (!setToDuplicate) {
-            return exercise;
-          }
-          const newSets = [
-            ...sets.slice(0, setIndex + 1),
-            { ...setToDuplicate, id: generateClientId() },
-            ...sets.slice(setIndex + 1),
-          ].map((s, i) => ({ ...s, set_number: i + 1 }));
-          return { ...exercise, sets: newSets };
-        })
-      );
-    },
-    []
-  );
-
-  const handleRemoveSet = useCallback(
-    (exerciseIndex: number, setIndex: number) => {
-      setExercises((prev) => {
-        const newState = prev.map((exercise, eIndex) => {
-          if (eIndex === exerciseIndex) {
-            const newSets = exercise.sets
-              .filter((_, sIndex) => sIndex !== setIndex)
-              .map((s, i) => ({ ...s, set_number: i + 1 }));
-            return { ...exercise, sets: newSets };
-          }
-          return exercise;
-        });
-        return newState.filter((exercise) => exercise.sets.length > 0);
-      });
-    },
-    []
-  );
-
-  const sensors = useSensors(
-    useSensor(PointerSensor),
-    useSensor(KeyboardSensor, {
-      coordinateGetter: sortableKeyboardCoordinates,
-    })
-  );
-
-  const handleDragEnd = (event: DragEndEvent) => {
-    const { active, over } = event;
-    if (!over || active.id === over.id) {
-      return;
-    }
-
-    const activeId = String(active.id);
-    const overId = String(over.id);
-
-    setExercises((prevExercises) => {
-      const newExercises = [...prevExercises];
-
-      // Try reordering exercises first
-      const activeExerciseIdx = newExercises.findIndex(
-        (ex) => String(ex.id) === activeId
-      );
-      const overExerciseIdx = newExercises.findIndex(
-        (ex) => String(ex.id) === overId
-      );
-
-      if (activeExerciseIdx !== -1 && overExerciseIdx !== -1) {
-        return arrayMove(newExercises, activeExerciseIdx, overExerciseIdx).map(
-          (ex, index) => ({
-            ...ex,
-            sort_order: index,
-          })
-        );
-      }
-
-      // If not exercises, check if it's a set reorder within the same exercise
-      const activeSetParentIdx = newExercises.findIndex((ex) =>
-        ex.sets.some((s) => String(s.id) === activeId)
-      );
-      const overSetParentIdx = newExercises.findIndex((ex) =>
-        ex.sets.some((s) => String(s.id) === overId)
-      );
-
-      if (
-        activeSetParentIdx !== -1 &&
-        overSetParentIdx !== -1 &&
-        activeSetParentIdx === overSetParentIdx
-      ) {
-        const exercise = newExercises[activeSetParentIdx];
-        if (!exercise) {
-          return prevExercises;
-        }
-        const oldSetIdx = exercise.sets.findIndex(
-          (s) => String(s.id) === activeId
-        );
-        const newSetIdx = exercise.sets.findIndex(
-          (s) => String(s.id) === overId
-        );
-
-        if (oldSetIdx !== -1 && newSetIdx !== -1) {
-          const reorderedSets = arrayMove(exercise.sets, oldSetIdx, newSetIdx);
-          newExercises[activeSetParentIdx] = {
-            ...exercise,
-            sets: reorderedSets.map((set, index) => ({
-              ...set,
-              set_number: index + 1,
-            })),
-          };
-        }
-      }
-
-      return newExercises;
-    });
-  };
-
-  const handleSubmit = () => {
-    if (!name.trim()) {
-      toast({
-        title: t('workoutPresetForm.validationErrorTitle', 'Validation Error'),
-        description: t(
-          'workoutPresetForm.nameRequiredError',
-          'Preset Name is required.'
-        ),
-        variant: 'destructive',
-      });
-      return;
-    }
-    debug(loggingLevel, 'WorkoutPresetForm: Submitting preset with data:', {
-      name,
-      description,
-      isPublic,
-      exercises,
-    });
-    onSave({
-      name,
-      description,
-      is_public: isPublic,
-      exercises: exercises.map((ex, index) => ({
-        ...ex,
-        sort_order: index,
-        sets: ex.sets.map((set) => ({
-          ...set,
-          weight: set.weight ?? 0, // already metric (kg) from UnitInput
-        })),
-      })),
-    });
-  };
+  const {
+    name,
+    description,
+    isPublic,
+    exercises,
+    isAddExerciseDialogOpen,
+    sensors,
+    setName,
+    setDescription,
+    setIsPublic,
+    setIsAddExerciseDialogOpen,
+    handleAddExercise,
+    handleRemoveExercise,
+    handleSetChange,
+    handleAddSet,
+    handleDuplicateSet,
+    handleRemoveSet,
+    handleDragEnd,
+    handleSubmit,
+  } = useWorkoutPresetForm({ onSave, initialPreset });
 
   return (
     <Dialog open={isOpen} onOpenChange={onClose}>
       <DialogContent
         requireConfirmation
-        className="sm:max-w-[1200px] max-h-[90vh] overflow-y-auto"
+        className="sm:max-w-300 max-h-[90vh] overflow-y-auto"
       >
         <DialogHeader>
           <DialogTitle>
@@ -741,7 +157,6 @@ const WorkoutPresetForm: React.FC<WorkoutPresetFormProps> = ({
                       handleRemoveSet={handleRemoveSet}
                       handleAddSet={handleAddSet}
                       weightUnit={weightUnit}
-                      t={t}
                     />
                   ))}
                 </div>
